pyproject.toml
[project]
name = "trials"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deep-translator>=1.11.4",
    "fastapi>=0.122.0",
    "langchain>=1.2.0",
    "langchain-google-genai>=4.1.2",
    "mangum>=0.19.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.38.0",
    "youtube-transcript-api>=1.2.3",
]
